rename the working compare models new to compare models
